feat: introduce StatusBadge component
This component can be used to render badges with additional information
as seen in the navbar tab buttons and menu items.

Here's how it can be used:

```
StatusBadge {
  value: 2
}
```

By default and based on `value`, `StatusBadge` will change its width.
If no `value` is provided, it renders as badge indicator as seen in the profile
tab button.

Closes #2453