feature(@desktop/chat): support jumping to search result message that is not currently loaded in memory
A crash leaving chat is fixed.

The issue was very hidden cause everything looks ok, at first glance, but not, the thing is that
"messages" exposed to qml were deleted from ChatsView and signals beginRemoveRows and
endRemoveRows were emitted for CatsView abstract list model instead for MessageView abstract list
model. That results in an app crash in some moments (not always reproducible, but often).

Fixes: #3005