fix(Communities): move backup community popup logic to onClosed handler
The transfer ownership popup triggered by the backup community baner button
was behaving weirdly because the banner from where it originates lives inside
a loader that is deactivated once the popup was opened.

Moving the logic that deactivates the loader into the onClosed handler of the
popup solves this issue.

Thanks to @pinebit for getting to the bottom of this!

Fixes #2997