fix(@desktop/browser): Favorite list get truncated when its too long
Adapted the bookmark list to a grid view to accomodate items when the list gets long.
Max column size is 7 and in case screen cannot show 7 columns it will flow to the next row.
Also added ideinticon for a website with no icon
Fixed the issue of DApps launching on all new tabs.

fixes #2009