refactor(@desktop/general): move Qt logic out of src/status library
Since `src/status` was Qt dependant part because of the following foundation files:
- /src/status/tasks/marathon/mailserver/controller.nim
- /src/status/signals/core.nim
- /src/status/tasks/marathon/mailserver/events.nim
- /src/status/tasks/qt.nim

and because logic related classes like:
- /src/status/chat.nim

this commit made that part Qt independant.

New layer `src/app_service` is introduced and above mentioned foundation files
are moved there. As well as corresponding logic for requested services.

Communication between logical parts on the high level is as follows:
- `src/app` => `src/app_service` => `src/status` => `src/app`
- `src/app` => `src/status` => `src/app`

Fixes: #3376